requirements.txt
blurhash==1.1.4
certifi==2020.6.20
chardet==3.0.4
decorator==5.1.1
feedparser==6.0.10
idna==2.10
Mastodon.py==1.5.1
oauthlib==3.1.0
pkg_resources==0.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
requests==2.24.0
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
urllib3==1.25.10
